ai-server/requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.26,<3
opencv-python>=4.8,<5
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0

# logging -------------------------------------
tensorboard>=2.4.1
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas
Pinyin2Hanzi>=0.1.1

# export --------------------------------------
# coremltools>=4.1
# onnx>=1.9.0
# scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
# Cython  # for pycocotools https://github.com/cocodataset/cocoapi/issues/172
# pycocotools>=2.0  # COCO mAP
# albumentations>=1.0.3
thop  # FLOPs computation

# mediapipe -----------------------------------
mediapipe>=0.10.14,<0.11

# server (FastAPI / WebRTC) -------------------
fastapi>=0.115,<1
uvicorn>=0.30,<1
python-multipart>=0.0.9,<1
aiortc>=1.9,<2
openai
python-dotenv
requests
